Fix #10526: Datatable Selenium add row toggler
